fix(action): ensure modules located at top level in site-packages zip archive
Prior to this commit, the zip file contained the site-packages directory at the
top level, whereas the goal was to have the contents of the site-packages
directory at the top level of the zip.

In this commit, the command is updated to first navigate into the site-packages
directory before creating the zip file, ensuring only the contents of the
directory are included at the top level of the archive.